requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
autopep8==2.3.1
black==24.8.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
django==4.2.23
django-admin-env-notice==1.0.1
django-admin-sortable2==2.2.8
django-cors-headers==4.4.0
django-csp==4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-hijack==3.6.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-prose-editor==0.16.1
django-silk==5.2.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
flake8==7.1.2
gprof2dot==2025.4.14
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langsmith==0.4.25
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.1.0
nh3==0.3.0
openai==1.106.1
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
pip==25.2
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyflakes==3.2.0
pyjwt==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.0.1
pytz==2025.2
pyyaml==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
ruff==0.12.1
setuptools==80.9.0
shortuuid==1.0.13
sniffio==1.3.1
sqlalchemy==2.0.43
sqlparse==0.5.3
stripe==12.5.1
tenacity==9.1.2
tiktoken==0.11.0
tomli==2.2.1
tqdm==4.67.1
types-pyyaml==6.0.12.20241230
typing-extensions==4.13.2
typing-inspection==0.4.1
uritemplate==4.1.1
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.7.0
zstandard==0.24.0
